% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasf-package.R
\docType{package}
\name{sasf-package}
\alias{sasf}
\alias{sasf-package}
\title{sasf: South African shapefiles for R users}
\description{
The \code{sasf} package makes spatial data for South Africa accessible to R users.
}
\author{
\strong{Maintainer}: Wihan Marais \email{wihanmarais@sun.ac.za} (\href{https://orcid.org/0000-0002-4720-0929}{ORCID}) [copyright holder]

}
\keyword{internal}
