% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_asp}
\alias{get_asp}
\title{Calculate aspect ratio and optimal dimensions for sf objects}
\usage{
get_asp(sf_obj, target_width = NULL)
}
\arguments{
\item{sf_obj}{An sf object with valid geometries}

\item{target_width}{Optional numeric value > 0 for desired plot width}
}
\value{
A list containing:
\item{asp}{The aspect ratio (width/height) of the bounding box, adjusted for latitude in geographic coordinates}
\item{target_width}{The specified target width (if provided)}
\item{target_height}{The calculated optimal height based on aspect ratio}
}
\description{
Calculates the aspect ratio of a simple features (sf) object and optionally
determines optimal plot dimensions based on a target width. For geographic
coordinates, the aspect ratio is adjusted for latitude distortion.
}
\author{
Wihan Marais <github.com/WihanZA>
}
